+++ lib/player/platform_player.dart
@@ -37,4 +37,41 @@ class PlatformPlayer extends SimplePlayer
         ) {
     setupJumper(maxJump: countJumps);
   }
+
+  @override
+  Future<void> replaceAnimation(
+    SimpleDirectionAnimation newAnimation, {
+    bool doIdle = false,
+  }) {
+    throw Exception(
+      'In PlatformPlayer update animation using replacePlatformAnimation',
+    );
+  }
+
+  Future<void> replacePlatformAnimation(
+    PlatformAnimations animation, {
+    bool doIdle = false,
+  }) {
+    return super.replaceAnimation(
+      SimpleDirectionAnimation(
+        idleRight: animation.idleRight,
+        runRight: animation.runRight,
+        idleLeft: animation.idleLeft,
+        runLeft: animation.runLeft,
+        others: {
+          if (animation.jump?.jumpUpRight != null)
+            JumpAnimationsEnum.jumpUpRight.name: animation.jump!.jumpUpRight,
+          if (animation.jump?.jumpUpLeft != null)
+            JumpAnimationsEnum.jumpUpLeft.name: animation.jump!.jumpUpLeft!,
+          if (animation.jump?.jumpDownRight != null)
+            JumpAnimationsEnum.jumpDownRight.name:
+                animation.jump!.jumpDownRight,
+          if (animation.jump?.jumpDownLeft != null)
+            JumpAnimationsEnum.jumpDownLeft.name: animation.jump!.jumpDownLeft!,
+          ...animation.others ?? {},
+        },
+      ),
+      doIdle: doIdle,
+    );
+  }
 }