+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - `GameObject` improvements
 - `GameDecoration` improvements
 - Adds `initPosition` param in `CameraConfig`
+- Adds `replacePlatformAnimation` method in `PlatformPlayer`
 
 # 3.0.0-beta.14